% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist.pt}
\alias{dist.pt}
\title{dist.pt}
\usage{
dist.pt(x, y, z)
}
\arguments{
\item{x}{int: the point}

\item{y}{int: the standard deviation of the 1st PC}

\item{zint:}{the standard deviation of the last PC}
}
\value{

}
\description{
Distance between a line and point
}
